+++ src/main/java/com/bookstore/online/domain/review/dto/request/PostReviewRequestDto.java
@@ -28,8 +28,4 @@ public class PostReviewRequestDto {
   @NotBlank
   private String comment;
 
-//  @NotNull
-//  @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-//  private LocalDateTime createdAt;
-
 }
